Inputs validation moved to custom hook use-validation.js
