+++ src/components/SignInForm/SignInForm.js
@@ -1,5 +1,7 @@
 import { useState } from 'react';
 
+import useValidation from '../../hooks/use-validation';
+
 import { Button } from '../UI/Button.styled';
 import { Form } from './SignInForm.styled';
 import Input from '../UI/Input';
@@ -8,19 +10,7 @@ const SignInForm = () => {
   const [enteredUsername, setEnteredUsername] = useState('');
   const [enteredPassword, setEnteredPassword] = useState('');
 
-  const isEmailValid = (email) => {
-    const emailVailid = /^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$/;
-
-    return emailVailid.test(email);
-  };
-
-  const isPasswordValid = (password) => {
-    const passwordValid = /^(?=.*[a-zA-Z])(?=.*[0-9])/;
-
-    return passwordValid.test(password);
-  };
-
-  const isLengthMoreEight = enteredPassword.trim().length > 7;
+  const isValid = useValidation(enteredUsername, enteredPassword);
 
   const enteredUsernameHandler = (event) => {
     setEnteredUsername(event.target.value);
@@ -28,15 +18,6 @@ const SignInForm = () => {
 
   const enteredPasswordHandler = (event) => {
     setEnteredPassword(event.target.value);
-  };
-
-  const isValid = () =>
-    isEmailValid(enteredUsername) &&
-    isLengthMoreEight &&
-    isPasswordValid(enteredPassword);
-
-  const onClickHandler = () => {
-    isValid();
   };
 
   const onSubmitHandler = (event) => {
@@ -59,9 +40,7 @@ const SignInForm = () => {
         placeholder="Enter your password"
         required
       />
-      <Button onClick={onClickHandler} disabled={!isValid()}>
-        Log in
-      </Button>
+      <Button disabled={!isValid()}>Log in</Button>
     </Form>
   );
 };
